refactor + skriver tester på onSubmit
