+++ src/components/UtenlandsoppholdListe/UtenlandsoppholdListe.module.scss.d.ts
@@ -1,0 +1,13 @@
+import globalClassNames from "../../style.d";
+declare const classNames: typeof globalClassNames & {
+  readonly section: "section";
+  readonly bodyshort: "bodyshort";
+  readonly utenlandsperioderItem: "utenlandsperioderItem";
+  readonly utenlandsperioder: "utenlandsperioder";
+  readonly utenlandsperioderButtons: "utenlandsperioderButtons";
+  readonly utenlandsperioderButtons__endre: "utenlandsperioderButtons__endre";
+  readonly utenlandsperioderButtons__slette: "utenlandsperioderButtons__slette";
+  readonly utenlandsperioderText: "utenlandsperioderText";
+  readonly error: "error";
+};
+export = classNames;
