+++ src/components/stegvisning/Utenlandsopphold/Utenlandsopphold.module.scss.d.ts
@@ -3,15 +3,6 @@ declare const classNames: typeof globalClassNames & {
   readonly readmore1: "readmore1";
   readonly readmore2: "readmore2";
   readonly radiogroup: "radiogroup";
-  readonly bodyshort: "bodyshort";
-  readonly section: "section";
-  readonly utenlandsperioderItem: "utenlandsperioderItem";
-  readonly utenlandsperioder: "utenlandsperioder";
-  readonly utenlandsperioderButtons: "utenlandsperioderButtons";
-  readonly utenlandsperioderButtons__endre: "utenlandsperioderButtons__endre";
-  readonly utenlandsperioderButtons__slette: "utenlandsperioderButtons__slette";
-  readonly utenlandsperioderText: "utenlandsperioderText";
   readonly button: "button";
-  readonly error: "error";
 };
 export = classNames;
