+++ src/translations/nn.ts
@@ -131,6 +131,9 @@ const translations = {
   'utenlandsopphold.om_oppholdet_ditt_modal.title': 'MANGLER_TEKST',
   'utenlandsopphold.om_oppholdet_ditt_modal.button.legg_til': 'MANGLER_TEKST',
   'utenlandsopphold.om_oppholdet_ditt_modal.button.oppdater': 'MANGLER_TEKST',
+  'utenlandsopphold.om_oppholdet_ditt_modal.button.avbryt': 'MANGLER_TEKST',
+  'utenlandsopphold.om_oppholdet_ditt_modal.button.avbryt_endring':
+    'MANGLER_TEKST',
   'utenlandsopphold.om_oppholdet_ditt_modal.land.label': 'MANGLER_TEKST',
   'utenlandsopphold.om_oppholdet_ditt_modal.har_jobbet.label': 'MANGLER_TEKST',
   'utenlandsopphold.om_oppholdet_ditt_modal.har_jobbet.description':
